Update README.md
This repository contains the resources pertaining to  Public and Patient Involvement and Engagement (PPIE), developed by the Data Analytics directorate within the Health Foundation. 

It includes our primary tool, a PPIE framework "Guide to Action", which is a practical guide to undertaking appropriate involvement and engagement activities in health care analytics. 

The accompanying Decision Tree document helps users decide upon the appropriate type of involvement and engagement for their project, and points them to the relevant section in the Guide to Action.